Strict one-component-per-file ontology.
This makes it a bit easier to see what a single component does, at the cost of
a slightly more sprawling filesystem.
