+++ src/app/components/Landing/Copy.js
@@ -1,0 +1,16 @@
+import React from 'react'
+
+export default function Copy() {
+  return <div className="jumbotron">
+    <h1>Distant Shore</h1>
+    <p>
+      Knausgaard taxidermy gochujang polaroid. Lo-fi truffaut schlitz,
+      gluten-free forage pickled biodiesel beard meggings keytar art party
+      stumptown slow-carb craft beer.
+    </p>
+    <p>
+      In the finished app, this should be a preview of a live fight, or a
+      recording.
+    </p>
+  </div>
+}
